+++ lib/src/configuration/better_player_data_source_type.dart
@@ -1,4 +1,5 @@
 ///Source types of video. Network type is used for videos that are hosted on
 ///the web service. File type is used for videos that will be read from
 /// mobile device. SRT type is used for SRT (Secure Reliable Transport) streaming.
-enum BetterPlayerDataSourceType { network, file, memory, srt }
+/// UDP type is used for UDP (User Datagram Protocol) streaming.
+enum BetterPlayerDataSourceType { network, file, memory, srt, udp }